+++ internal/drapi/get_test.go
@@ -13,84 +13,3 @@
 // limitations under the License.
 
 package drapi
-
-import (
-	"context"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/datarobot/cli/internal/config"
-	"github.com/datarobot/cli/internal/config/viperx"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestGetUserID_Success(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		assert.Equal(t, "/api/v2/account/info/", r.URL.Path)
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte(`{"uid":"test-uid-123","email":"user@example.com"}`))
-	}))
-	defer server.Close()
-
-	defer resetTokenForTest(t, "test-token")()
-	defer viperx.Reset()
-
-	viperx.Set(config.DataRobotURL, server.URL+"/api/v2")
-
-	uid, err := GetUserID(context.Background())
-	require.NoError(t, err)
-	assert.Equal(t, "test-uid-123", uid)
-}
-
-func TestGetUserID_Unauthorized(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusUnauthorized)
-	}))
-	defer server.Close()
-
-	defer resetTokenForTest(t, "test-token")()
-	defer viperx.Reset()
-
-	viperx.Set(config.DataRobotURL, server.URL+"/api/v2")
-
-	uid, err := GetUserID(context.Background())
-	require.Error(t, err)
-	assert.Empty(t, uid)
-}
-
-func TestGetUserID_EmptyUID(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		assert.Equal(t, "/api/v2/account/info/", r.URL.Path)
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte(`{"uid":"","email":"user@example.com"}`))
-	}))
-	defer server.Close()
-
-	defer resetTokenForTest(t, "test-token")()
-	defer viperx.Reset()
-
-	viperx.Set(config.DataRobotURL, server.URL+"/api/v2")
-
-	uid, err := GetUserID(context.Background())
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "empty uid")
-	assert.Empty(t, uid)
-}
-
-func TestGetUserID_NetworkError(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
-		// Should never be called because we close the server before making the request.
-	}))
-	server.Close()
-
-	defer resetTokenForTest(t, "test-token")()
-	defer viperx.Reset()
-
-	viperx.Set(config.DataRobotURL, server.URL+"/api/v2")
-
-	uid, err := GetUserID(context.Background())
-	require.Error(t, err)
-	assert.Empty(t, uid)
-}
