+++ internal/telemetry/properties.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/datarobot/cli/internal/config"
 	"github.com/datarobot/cli/internal/config/viperx"
-	"github.com/datarobot/cli/internal/drapi"
+	"github.com/datarobot/cli/internal/userid"
 	"github.com/datarobot/cli/internal/version"
 )
 
@@ -66,7 +66,7 @@ func CollectCommonProperties() *CommonProperties {
 		}
 	}
 
-	rawUserID, _ := drapi.GetUserID(context.Background())
+	rawUserID, _ := userid.GetUserID(context.Background())
 	props.UserID = getOrCreateUserID(rawUserID)
 
 	return props
